add support for conditional requirements
Both tables and columns can have conditional requirements in the
data model, either 'CONDITIONAL (table_name)' for tables or
`CONDITIONAL (column = value)' for columns. Added a new workflow
based on data_dictionary_report, that takes the analysis TSV file
as an additional input to check conditional requirements for GSR
data files that are based on values of columns in the analysis
table. Other workflows just require updating the docker image
for AnvilDataModels_v0.1.1.